Rename BondsPool to MarketPreview. Remove unused BondModal component
