--- src/pages/BondsPoolPage/components/BondModal/BondModal.module.scss
@@ -1,111 +1,0 @@
-.modal {
-  :global {
-    z-index: 2;
-
-    .ant-modal-body {
-      padding: 32px;
-    }
-
-    @media screen and (max-width: 767px) {
-      .ant-modal-body {
-        padding: 24px;
-      }
-    }
-  }
-  &::before {
-    content: '';
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    border: 1px solid var(--primary-text);
-    background-color: var(--primary-text);
-    top: 4px;
-    left: 4px;
-  }
-}
-
-.wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 4px;
-}
-
-.value {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  color: var(--secondary-text);
-  font: var(--font-subtitle);
-
-  svg {
-    width: 20px;
-  }
-}
-
-.title {
-  color: var(--secondary-border);
-  font: var(--font-text-6);
-}
-
-.closeModalSection {
-  height: 40px;
-  display: flex;
-  align-items: center;
-  background: var(--primary-background);
-  border-radius: 30px;
-  position: absolute;
-  top: -64px;
-  right: 0;
-
-  @media screen and (max-width: 1023px) {
-    height: 32px;
-    top: -48px;
-  }
-
-  @media screen and (max-width: 767px) {
-    height: 24px;
-    top: -38px;
-  }
-}
-.closeModalIcon {
-  height: 40px;
-  width: 40px;
-  border: 1px solid var(--primary-border);
-  background: var(--primary-background);
-  border-radius: 50%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-right: -1px;
-  cursor: pointer;
-
-  @media screen and (max-width: 1024px) {
-    height: 32px;
-    width: 32px;
-  }
-  @media screen and (max-width: 767px) {
-    height: 24px;
-    width: 24px;
-  }
-}
-
-.closeIcon {
-  width: 12px;
-
-  stroke: var(--secondary-border);
-
-  rect {
-    fill: var(--secondary-border);
-  }
-
-  @media screen and (max-width: 1024px) {
-    width: 10px;
-  }
-}
