--- src/pages/BondsPoolPage/components/BondModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BondModal';
