--- src/pages/BondsPoolPage/components/BondPool/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BondPool';
