+++ src/pages/MarketsPreviewPage/MarketsPreviewPage.module.scss
