+++ src/pages/MarketsPreviewPage/components/MarketPreviewCard/MarketPreviewCard.module.scss
