+++ src/pages/MarketsPreviewPage/components/MarketPreviewCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MarketPreviewCard';
