+++ src/pages/MarketsPreviewPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MarketsPreviewPage';
