+++ src/pages/index.ts
@@ -4,6 +4,6 @@ export { default as LendPage } from './LendPage';
 export { default as LiquidationsPage } from './LiquidationsPage';
 export { default as LoansPage } from './LoansPage';
 export { default as BorrowPage } from './BorrowPage';
-export { default as BondsPoolPage } from './BondsPoolPage';
+export { default as MarketsPreviewPage } from './MarketsPreviewPage';
 export { default as BondPage } from './BondPage';
 export { default as PoolsCreationPage } from './PoolsCreationPage';
